backend-python/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
google-api-python-client==2.152.0
google-auth==2.34.0
google-auth-httplib2==0.2.0
python-dotenv==1.0.1
pydantic==2.9.2
python-multipart==0.0.12

